app/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.6
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.29.0
sniffio==1.3.0
starlette==0.26.1
tinify==1.6.0
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
yarl==1.9.2
